Git-Repo-Analysis/requirements.txt
gitpython
python-dotenv
google-generativeai
fastapi
uvicorn
pydantic
streamlit
